move ClientFactoryException
