+++ aws-shared/src/main/java/com/kdgregory/logging/aws/common/DefaultClientFactory.java
@@ -25,6 +25,7 @@ import com.amazonaws.regions.Regions;
 import com.kdgregory.logging.aws.internal.ReflectionBasedInvoker;
 import com.kdgregory.logging.aws.internal.retrievers.RoleArnRetriever;
 import com.kdgregory.logging.common.factories.ClientFactory;
+import com.kdgregory.logging.common.factories.ClientFactoryException;
 import com.kdgregory.logging.common.util.InternalLogger;
