+++ aws-shared/src/main/java/com/kdgregory/logging/common/factories/ClientFactoryException.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.kdgregory.logging.aws.common;
+package com.kdgregory.logging.common.factories;
 
 import java.lang.reflect.InvocationTargetException;
